refactor(spec): move qwen35 DFlash state + DflashSpeculator into the qwen crate
Step 8 of the spec-accept-core/DFlash-relocation plan. `DdtreeState`,
`DflashState`, `load_dflash_state`, `lower_qwen35`, `DflashSpeculator`, and
`build_dflash_speculator` move from `hipfire-loader` into a new
`hipfire_arch_qwen35::dflash_spec` module — all qwen35 + runtime types, no
loader types, so the qwen35 crate now owns its DFlash mechanics end-to-end and
the loader only constructs/routes them.

Pure relocation: same functions, same call order, byte-identical behavior. The
loader keeps `Qwen35SlotGuard` + the `build_speculator` registry (which now
dispatches to the qwen35 `build_dflash_speculator`); its load path calls
`dflash_spec::load_dflash_state`. The daemon is untouched (still names
`Qwen35SlotGuard` via the `SpecSlotGuard` enum — that becomes generic in step 9).

Verified: builds (loader + daemon example), 6 accept-core unit tests pass,
rustfmt-clean (per-file skip_children — NOT fmt-changed, which churns this
branch's legacy debt), clippy-clean. The 27B DFlash serve-multiturn behavioral
gate is deferred to step 9, where the daemon guard dispatch actually changes.
